py/requirements.txt
supabase>=2.4.0
python-dotenv
requests